permissões
adição de permissões de usuário e admin.

+++ app.py
@@ -46,7 +46,7 @@ def create_user():
         password = data.get("password")
 
         if username and password:
-                user = User(username=username, password=password)
+                user = User(username=username, password=password, role='user')
                 db.session.add(user)
                 db.session.commit()
                 return jsonify({"message": "Usuário cadastrado com sucesso."})
@@ -68,7 +68,9 @@ def read_user(id_user):
 def update_user(id_user):
         data = request.json
         user = User.query.get(id_user)
-
+        if id_user != current_user.id and current_user.role == "user":
+                return jsonify({"message": "Operação não permitida"}), 403
+        
         if user and data.get("password"):
                 user.password = data.get("password")
                 db.session.commit()
@@ -82,8 +84,11 @@ def update_user(id_user):
 def delete_user(id_user):
         user = User.query.get(id_user)
 
+        if current_user.role != 'admin':
+                return jsonify({"message": "operação não permitida."})
+
         if id_user == current_user.id:
-                return jsonify({"message": "Delção não permitida."}), 403
+                return jsonify({"message": "Deleção não permitida."}), 403
 
         if user:
                 db.session.delete(user)
